excluding admin from leaderboard

+++ pong-game/backend/app/game_service/models.py
@@ -232,9 +232,9 @@ class LeaderBoardManager(models.Manager):
 			return False
 
 	def updateLeaderBoard(self):
-		# if not self._hasGameOccured():
-			# return
-		players = CustomUser.objects.order_by('-mmr', '-winCount') # first by mmr and then by wins
+		if not self._hasGameOccured():
+			return
+		players = CustomUser.objects.filter(is_admin=False).order_by('-mmr', '-winCount') # first by mmr and then by wins
 
 		with transaction.atomic(): # blocks all other write operations for db & if this code block fails, rewinds to before function call
 			self.model.objects.all().delete()
